+++ development/metamaskbot-build-announce.js
@@ -55,6 +55,9 @@ async function start() {
     })
     .join(', ')
 
+  const coverageUrl = `${BUILD_LINK_BASE}/coverage/index.html`
+  const coverageLink = `<a href="${coverageUrl}">Report</a>`
+
   // links to bundle browser builds
   const depVizUrl = `${BUILD_LINK_BASE}/build-artifacts/deps-viz/background/index.html`
   const depVizLink = `<a href="${depVizUrl}">background</a>`
@@ -65,6 +68,7 @@ async function start() {
   const contentRows = [
     `builds: ${buildLinks}`,
     `bundle viz: ${bundleLinks}`,
+    `code coverage: ${coverageLink}`,
     `dep viz: ${depVizLink}`,
     `<a href="${allArtifactsUrl}">all artifacts</a>`,
   ]