+++ test/unit/app/controllers/permissions/permissions-middleware-test.js
@@ -18,6 +18,20 @@ const { CAVEATS, ERRORS, PERMS, RPC_REQUESTS } = getters
 
 const { ACCOUNTS, DOMAINS, PERM_NAMES } = constants
 
+const initPermController = () => {
+  return new PermissionsController({
+    ...getPermControllerOpts(),
+  })
+}
+
+const createApprovalSpies = (permController) => {
+  sinon.spy(permController.approvals, '_add')
+}
+
+const getNextApprovalId = (permController) => {
+  return permController.approvals._approvals.keys().next().value
+}
+
 const validatePermission = (perm, name, origin, caveats) => {
   assert.equal(
     name,
@@ -36,12 +50,6 @@ const validatePermission = (perm, name, origin, caveats) => {
   }
 }
 
-const initPermController = () => {
-  return new PermissionsController({
-    ...getPermControllerOpts(),
-  })
-}
-
 describe('permissions middleware', function () {
   describe('wallet_requestPermissions', function () {
     let permController
@@ -52,6 +60,8 @@ describe('permissions middleware', function () {
     })
 
     it('grants permissions on user approval', async function () {
+      createApprovalSpies(permController)
+
       const aMiddleware = getPermissionsMiddleware(
         permController,
         DOMAINS.a.origin,
@@ -72,13 +82,12 @@ describe('permissions middleware', function () {
 
       await userApprovalPromise
 
-      assert.equal(
-        permController.pendingApprovals.size,
-        1,
-        'perm controller should have single pending approval',
+      assert.ok(
+        permController.approvals._add.calledOnce,
+        'should have added single approval request',
       )
 
-      const id = permController.pendingApprovals.keys().next().value
+      const id = getNextApprovalId(permController)
       const approvedReq = PERMS.approvedRequest(
         id,
         PERMS.requests.eth_accounts(),
@@ -150,7 +159,7 @@ describe('permissions middleware', function () {
 
       await userApprovalPromise
 
-      const id1 = permController.pendingApprovals.keys().next().value
+      const id1 = getNextApprovalId(permController)
       const approvedReq1 = PERMS.approvedRequest(
         id1,
         PERMS.requests.eth_accounts(),
@@ -219,7 +228,7 @@ describe('permissions middleware', function () {
 
       await userApprovalPromise
 
-      const id2 = permController.pendingApprovals.keys().next().value
+      const id2 = getNextApprovalId(permController)
       const approvedReq2 = PERMS.approvedRequest(id2, { ...requestedPerms2 })
 
       await permController.approvePermissionsRequest(
@@ -275,6 +284,8 @@ describe('permissions middleware', function () {
     })
 
     it('rejects permissions on user rejection', async function () {
+      createApprovalSpies(permController)
+
       const aMiddleware = getPermissionsMiddleware(
         permController,
         DOMAINS.a.origin,
@@ -298,13 +309,12 @@ describe('permissions middleware', function () {
 
       await userApprovalPromise
 
-      assert.equal(
-        permController.pendingApprovals.size,
-        1,
-        'perm controller should have single pending approval',
+      assert.ok(
+        permController.approvals._add.calledOnce,
+        'should have added single approval request',
       )
 
-      const id = permController.pendingApprovals.keys().next().value
+      const id = getNextApprovalId(permController)
 
       await permController.rejectPermissionsRequest(id)
       await requestRejection
@@ -328,6 +338,8 @@ describe('permissions middleware', function () {
     })
 
     it('rejects requests with unknown permissions', async function () {
+      createApprovalSpies(permController)
+
       const aMiddleware = getPermissionsMiddleware(
         permController,
         DOMAINS.a.origin,
@@ -349,10 +361,9 @@ describe('permissions middleware', function () {
         'request should be rejected with correct error',
       )
 
-      assert.equal(
-        permController.pendingApprovals.size,
-        0,
-        'perm controller should have no pending approvals',
+      assert.ok(
+        permController.approvals._add.notCalled,
+        'no approval requests should have been added',
       )
 
       assert.ok(
@@ -367,7 +378,7 @@ describe('permissions middleware', function () {
     })
 
     it('accepts only a single pending permissions request per origin', async function () {
-      const expectedError = ERRORS.pendingApprovals.requestAlreadyPending()
+      createApprovalSpies(permController)
 
       // two middlewares for two origins
 
@@ -414,10 +425,9 @@ describe('permissions middleware', function () {
 
       await userApprovalPromise
 
-      assert.equal(
-        permController.pendingApprovals.size,
-        2,
-        'perm controller should have expected number of pending approvals',
+      assert.ok(
+        permController.approvals._add.calledTwice,
+        'should have added two approval requests',
       )
 
       // create and start processing second request for first origin,
@@ -430,6 +440,10 @@ describe('permissions middleware', function () {
       const resA2 = {}
 
       userApprovalPromise = getUserApprovalPromise(permController)
+
+      const expectedError = ERRORS.pendingApprovals.requestAlreadyPending(
+        DOMAINS.a.origin,
+      )
 
       const requestApprovalFail = assert.rejects(
         aMiddleware(reqA2, resA2),
@@ -447,17 +461,20 @@ describe('permissions middleware', function () {
         'response should have expected error and no result',
       )
 
-      // first requests for both origins should remain
-
       assert.equal(
-        permController.pendingApprovals.size,
+        permController.approvals._add.callCount,
+        3,
+        'should have attempted to create three pending approvals',
+      )
+      assert.equal(
+        permController.approvals._approvals.size,
         2,
-        'perm controller should have expected number of pending approvals',
+        'should only have created two pending approvals',
       )
 
       // now, remaining pending requests should be approved without issue
 
-      for (const id of permController.pendingApprovals.keys()) {
+      for (const id of permController.approvals._approvals.keys()) {
         await permController.approvePermissionsRequest(
           PERMS.approvedRequest(id, PERMS.requests.test_method()),
         )
@@ -483,12 +500,6 @@ describe('permissions middleware', function () {
         resB1.result.length,
         1,
         'second origin should have single approved permission',
-      )
-
-      assert.equal(
-        permController.pendingApprovals.size,
-        0,
-        'perm controller should have expected number of pending approvals',
       )
     })
   })
@@ -609,6 +620,8 @@ describe('permissions middleware', function () {
     })
 
     it('requests accounts for unpermitted origin, and approves on user approval', async function () {
+      createApprovalSpies(permController)
+
       const userApprovalPromise = getUserApprovalPromise(permController)
 
       const aMiddleware = getPermissionsMiddleware(
@@ -626,13 +639,12 @@ describe('permissions middleware', function () {
 
       await userApprovalPromise
 
-      assert.equal(
-        permController.pendingApprovals.size,
-        1,
-        'perm controller should have single pending approval',
+      assert.ok(
+        permController.approvals._add.calledOnce,
+        'should have added single approval request',
       )
 
-      const id = permController.pendingApprovals.keys().next().value
+      const id = getNextApprovalId(permController)
       const approvedReq = PERMS.approvedRequest(
         id,
         PERMS.requests.eth_accounts(),
@@ -685,6 +697,8 @@ describe('permissions middleware', function () {
     })
 
     it('requests accounts for unpermitted origin, and rejects on user rejection', async function () {
+      createApprovalSpies(permController)
+
       const userApprovalPromise = getUserApprovalPromise(permController)
 
       const aMiddleware = getPermissionsMiddleware(
@@ -705,13 +719,12 @@ describe('permissions middleware', function () {
 
       await userApprovalPromise
 
-      assert.equal(
-        permController.pendingApprovals.size,
-        1,
-        'perm controller should have single pending approval',
+      assert.ok(
+        permController.approvals._add.calledOnce,
+        'should have added single approval request',
       )
 
-      const id = permController.pendingApprovals.keys().next().value
+      const id = getNextApprovalId(permController)
 
       await permController.rejectPermissionsRequest(id)
       await requestRejection
@@ -788,7 +801,7 @@ describe('permissions middleware', function () {
       // this will reject because of the already pending request
       await assert.rejects(
         cMiddleware({ ...req }, {}),
-        ERRORS.eth_requestAccounts.requestAlreadyPending(),
+        ERRORS.eth_requestAccounts.requestAlreadyPending(DOMAINS.c.origin),
       )
 
       // now unlock and let through the first request
@@ -808,7 +821,7 @@ describe('permissions middleware', function () {
     })
   })
 
-  describe('wallet_sendDomainMetadata', function () {
+  describe('metamask_sendDomainMetadata', function () {
     let permController, clock
 
     beforeEach(function () {
@@ -828,7 +841,10 @@ describe('permissions middleware', function () {
         DOMAINS.c.origin,
       )
 
-      const req = RPC_REQUESTS.wallet_sendDomainMetadata(DOMAINS.c.origin, name)
+      const req = RPC_REQUESTS.metamask_sendDomainMetadata(
+        DOMAINS.c.origin,
+        name,
+      )
       const res = {}
 
       await assert.doesNotReject(cMiddleware(req, res), 'should not reject')
@@ -861,7 +877,10 @@ describe('permissions middleware', function () {
         extensionId,
       )
 
-      const req = RPC_REQUESTS.wallet_sendDomainMetadata(DOMAINS.c.origin, name)
+      const req = RPC_REQUESTS.metamask_sendDomainMetadata(
+        DOMAINS.c.origin,
+        name,
+      )
       const res = {}
 
       await assert.doesNotReject(cMiddleware(req, res), 'should not reject')
@@ -885,7 +904,10 @@ describe('permissions middleware', function () {
         DOMAINS.c.origin,
       )
 
-      const req = RPC_REQUESTS.wallet_sendDomainMetadata(DOMAINS.c.origin, name)
+      const req = RPC_REQUESTS.metamask_sendDomainMetadata(
+        DOMAINS.c.origin,
+        name,
+      )
       const res = {}
 
       await assert.doesNotReject(cMiddleware(req, res), 'should not reject')
@@ -907,7 +929,7 @@ describe('permissions middleware', function () {
         DOMAINS.c.origin,
       )
 
-      const req = RPC_REQUESTS.wallet_sendDomainMetadata(DOMAINS.c.origin)
+      const req = RPC_REQUESTS.metamask_sendDomainMetadata(DOMAINS.c.origin)
       delete req.domainMetadata
       const res = {}